Move torch.compile configs into cuda_graph_runner.py (#993)
